Restructuring the Search Field.
Making some minor tweaks to the Search field and refactor to use Search
API generic name conventions.
